+++ win32/gui.py
@@ -37,7 +37,7 @@ import yaml
 EXCLUDED_WINDOWS_CHECKS = [
     'cacti', 'directory', 'gearmand',
     'hdfs', 'kafka_consumer', 'mcache', 'network',
-    'redis', 'postfix', 'process', 'gunicorn', 'zk',
+    'postfix', 'process', 'gunicorn', 'zk',
     ]
 
 MAIN_WINDOW_TITLE = "Datadog Agent Manager"